Update component catalog
